+++ engines/scumm/charset.cpp
@@ -22,8 +22,8 @@
 #include "graphics/font.h"
 #include "scumm/charset.h"
 #include "scumm/file.h"
-#include "scumm/gfx_mac.h"
 #include "scumm/scumm.h"
+#include "scumm/macgui/macgui.h"
 #include "scumm/nut_renderer.h"
 #include "scumm/util.h"
 #include "scumm/he/intern_he.h"
