+++ engines/scumm/cursor.cpp
@@ -30,8 +30,8 @@
 #include "scumm/charset.h"
 #include "scumm/he/intern_he.h"
 #include "scumm/object.h"
-#include "scumm/gfx_mac.h"
 #include "scumm/he/resource_he.h"
+#include "scumm/macgui/macgui.h"
 #include "scumm/scumm.h"
 #include "scumm/scumm_v2.h"
 #include "scumm/scumm_v5.h"
