+++ engines/scumm/macgui/macgui.h
