+++ engines/scumm/scumm.cpp
@@ -46,7 +46,6 @@
 #include "scumm/dialogs.h"
 #include "scumm/file.h"
 #include "scumm/file_nes.h"
-#include "scumm/gfx_mac.h"
 #include "scumm/imuse/imuse.h"
 #include "scumm/imuse_digi/dimuse_engine.h"
 #include "scumm/smush/smush_player.h"
@@ -57,6 +56,7 @@
 #include "scumm/he/logic_he.h"
 #include "scumm/he/sound_he.h"
 #include "scumm/object.h"
+#include "scumm/macgui/macgui.h"
 #include "scumm/players/player_ad.h"
 #include "scumm/players/player_nes.h"
 #include "scumm/players/player_sid.h"
